+++ src/main/java/subway/line/controller/LineController.java
@@ -1,11 +1,18 @@
-package subway.line;
+package subway.line.controller;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.station.Station;
-import subway.station.StationDao;
-import subway.station.StationResponse;
+import subway.line.domain.Line;
+import subway.line.dao.LineDao;
+import subway.line.dto.LineRequest;
+import subway.line.dto.LineResponse;
+import subway.section.domain.Section;
+import subway.section.dao.SectionDao;
+import subway.section.dto.SectionRequest;
+import subway.section.domain.Sections;
+import subway.station.domain.Station;
+import subway.station.dao.StationDao;
+import subway.station.dto.StationResponse;
 
 import java.net.URI;
 import java.util.stream.Collectors;
