+++ src/main/java/subway/line/domain/Line.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.domain;
 
 public class Line {
     public static final long HEAD = 0;
