+++ src/main/java/subway/section/dao/SectionDao.java
@@ -1,10 +1,11 @@
-package subway.line;
+package subway.section.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import subway.section.domain.Section;
 
 import java.sql.PreparedStatement;
 import java.util.List;
