+++ src/main/java/subway/section/domain/Sections.java
@@ -1,4 +1,6 @@
-package subway.line;
+package subway.section.domain;
+
+import subway.section.domain.Section;
 
 import java.util.List;
 import java.util.stream.Collectors;
