+++ src/main/java/subway/station/controller/StationController.java
@@ -1,8 +1,12 @@
-package subway.station;
+package subway.station.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.station.domain.Station;
+import subway.station.dao.StationDao;
+import subway.station.dto.StationRequest;
+import subway.station.dto.StationResponse;
 
 import java.net.URI;
 import java.util.List;
