+++ src/main/java/subway/station/dto/StationResponse.java
@@ -1,4 +1,6 @@
-package subway.station;
+package subway.station.dto;
+
+import subway.station.domain.Station;
 
 public class StationResponse {
     private Long id;
